requirements.txt
numpy
matplotlib
pandas
homeassistant-api
requests
PyYAML
Pillow
pytz
cairosvg